Fix transaction on delete

+++ app/repos/search-match-codes.js
@@ -42,8 +42,10 @@ const insertMatchCode = async (personId, fieldValue, simple, transaction) => {
 }
 
 const updateMatchCodesPerPerson = async (personId, row, transaction) => {
-  await sequelize.models.search_match_code.destroy({ where: { person_id: personId }, force: true }, { transaction })
-  await insertPersonMatchCodes(row)
+  if (personId) {
+    await sequelize.models.search_match_code.destroy({ where: { person_id: personId }, force: true }, { transaction })
+    await insertPersonMatchCodes(row, transaction)
+  }
 }
 
 const buildFuzzyCodes = (terms) => {
@@ -57,28 +59,7 @@ const buildFuzzyCodes = (terms) => {
 
 const fuzzySearch = async (terms) => {
   const fuzzyCodes = buildFuzzyCodes(terms)
-  /*
-  const fuzzyResults = await sequelize.models.search_match_code.findAll({
-    attributes: ['person_id', [sequelize.fn('count', sequelize.col('search_match_code.person_id')), 'matchcount']],
-    where: {
-      match_code: fuzzyCodes
-    },
-    group: ['person_id'],
-    order: [['matchcount', 'DESC']],
-    raw: true
-  })
-  const numResults = fuzzyResults.length
-  const uniquePersons = []
-  if (numResults > 0) {
-    const highestMatch = parseInt(fuzzyResults[0].matchcount)
-    console.log('high', highestMatch)
-    const matchThreshold = 1 // highestMatch / 4
-    fuzzyResults.forEach(res => {
-      if (parseInt(res.matchcount) >= matchThreshold && !uniquePersons.includes(res.person_id)) {
-        uniquePersons.push(res.person_id)
-      }
-    })
-  */
+
   const fuzzyResults = await sequelize.models.search_match_code.findAll({
     where: {
       match_code: fuzzyCodes
